requirements.txt
pandas==2.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0
scipy==1.11.3 